kern/vfs: Add the per-filesystem vnode counter to struct mount.
This is the preparation for the precise vnode recycle on ZFS and nullfs.
In the upcoming commits, those filesystems make mnt_fsvninusep point to the
vnode counter internal in each filesystem.

This change does not affect any filesystems that do not set mnt_fsvninusep.

This change is separated from the ZFS and nullfs fixes because it modifies
struct mount.

PR: 275594
Reported By: Thomas Mueller <thmu7@freenet.de>
Security: FreeBSD-EN-23:18.openzfs
Signed-off-by: Seigo Tanimura <seigo.tanimura@gmail.com>